requirements.txt
tensorflow==2.21.0
numpy==2.4.3
matplotlib==3.10.8
seaborn==0.13.2
scikit-learn==1.8.0
pillow==12.1.1
pydot==4.0.1
tf2onnx==1.17.0
onnxruntime==1.24.4
